OM-345: Determine if no-local-state can be preserved in React 0.13
It appears it can be done at least as well as the behavior we had before.
This commit simply delays recording state until after mounting.

A further enhancement would be to use the normal state mechanism
right up until mounting at which the global state map is used. Likely
this is all that's need to get the CircleCI level of state re-hydration.